requirements.txt
aioboto3==10.3.0 ; python_version >= "3.8" and python_version < "4.0"
aiobotocore[boto3]==2.4.1 ; python_version >= "3.8" and python_version < "4.0"
aiohttp==3.7.4.post0 ; python_version >= "3.8" and python_version < "4.0"
aioitertools==0.11.0 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==3.0.1 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.8" and python_version < "4.0"
boto3==1.24.59 ; python_version >= "3.8" and python_version < "4.0"
botocore==1.27.59 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.3.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.8" and python_version < "4.0"
chardet==4.0.0 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.8" and python_version < "4"
click==8.1.3 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
cryptography==39.0.0 ; python_version >= "3.8" and python_version < "4.0"
enum-compat==0.0.3 ; python_version >= "3.8" and python_version < "4.0"
flask
google-api-core==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
google-api-python-client==2.79.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth-httplib2==0.1.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth-oauthlib==1.0.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth==2.16.1 ; python_version >= "3.8" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version >= "3.8" and python_version < "4.0"
grpcio
gunicorn==20.1.0 ; python_version >= "3.8" and python_version < "4.0"
httplib2==0.21.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4"
importlib-metadata==6.1.0 ; python_version >= "3.8" and python_version < "3.10"
itsdangerous==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.1 ; python_version < "4.0" and python_version >= "3.8"
oauthlib==3.2.2 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.8" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.8" and python_version < "4.0"
prometheus-client==0.16.0 ; python_version >= "3.8" and python_version < "4.0"
protobuf==3.20.3 ; python_version >= "3.8" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.8" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.8" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.8" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.8" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.8" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "4.0"
pytelegrambotapi==4.9.0 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.8" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.8" and python_version < "4"
rsa==4.9 ; python_version >= "3.8" and python_version < "4"
s3transfer==0.6.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==67.7.2 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
telebot==0.0.5 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.8" and python_version < "4"
werkzeug
wrapt==1.14.1 ; python_version >= "3.8" and python_version < "4.0"
yadisk==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
yandexcloud==0.194.0 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
ydb==2.12.1 ; python_version >= "3.8" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.8" and python_version < "3.10"
